+++ src/Core/Kernel.php
@@ -43,5 +43,9 @@ class Kernel {
         if (!empty($mods['checkout']) && class_exists('HB\\UCS\\Modules\\Checkout\\CheckoutModule')) {
             (new \HB\UCS\Modules\Checkout\CheckoutModule())->init();
         }
+
+        if (!empty($mods['product_pages']) && class_exists('HB\\UCS\\Modules\\ProductPages\\ProductPagesModule')) {
+            (new \HB\UCS\Modules\ProductPages\ProductPagesModule())->init();
+        }
     }
 }